# apps/backend/openapi.json
{
	"openapi": "3.0.3",
	"info": {
		"description": "API's used and provided by [https://StellarAtlas.io](https://StellarAtlas.io).",
		"title": "StellarAtlas.io API",
		"contact": {
			"name": "StellarAtlas",
			"email": "info@StellarAtlas.io"
		},
		"version": "v1"
	},
	"servers": [
		{
			"url": "https://api.StellarAtlas.io",
			"description": "API for Stellar public network"
		},
		{
			"url": "https://api-testnet.StellarAtlas.io",
			"description": "API for Stellar testnet"
		}
	],
...
	"paths": {
...
		"/v1/status/full-history": {
			"get": {
				"summary": "Get full-history parser status",
				"description": "Returns persisted full-history parser readiness from archive-derived parsed ledger headers.",
				"tags": ["Status", "Explorer"],
				"operationId": "getFullHistoryStatus",
				"responses": {
					"200": {
						"description": "Full-history status",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/FullHistoryStatus" }
							}
						}
					},
					"500": { "description": "Internal server error" }
				}
			}
		},
		"/v1/status/ingestion": {
			"get": {
				"summary": "Get full-history ingestion status",
				"description": "Returns parser status with archive indexing job queue summary.",
				"tags": ["Status", "Explorer"],
				"operationId": "getIngestionStatus",
				"responses": {
					"200": {
						"description": "Ingestion status",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/IngestionStatus" }
							}
						}
					},
					"500": { "description": "Internal server error" }
				}
			}
		},
		"/v1/indexing/jobs": {
			"get": {
				"summary": "List recent indexing jobs",
				"description": "Returns bounded recent archive indexing jobs from the persisted scanner queue.",
				"tags": ["Explorer", "Status"],
				"operationId": "getIndexingJobs",
				"parameters": [
					{
						"name": "limit",
						"in": "query",
						"required": false,
						"schema": { "type": "integer", "minimum": 1, "maximum": 100 }
					}
				],
				"responses": {
					"200": {
						"description": "Recent indexing jobs",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/IndexingJobs" }
							}
						}
					},
					"400": { "description": "Invalid limit" },
					"500": { "description": "Internal server error" }
				}
			}
		},
		"/v1/indexing/ranges": {
			"get": {
				"summary": "List parsed archive ranges",
				"description": "Returns per-archive parsed ledger header coverage from persisted parser evidence.",
				"tags": ["Explorer", "Status"],
				"operationId": "getIndexingRanges",
				"parameters": [
					{
						"name": "limit",
						"in": "query",
						"required": false,
						"schema": { "type": "integer", "minimum": 1, "maximum": 100 }
					}
				],
				"responses": {
					"200": {
						"description": "Parsed archive ranges",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/IndexingRanges" }
							}
						}
					},
					"400": { "description": "Invalid limit" },
					"500": { "description": "Internal server error" }
				}
			}
		},
		"/v1/ledgers/{sequence}/ingestion-status": {
			"get": {
				"summary": "Get ledger ingestion status",
				"description": "Reports whether an archive-derived parsed ledger header is available for a ledger sequence.",
				"tags": ["Explorer", "Status"],
				"operationId": "getLedgerIngestionStatus",
				"parameters": [
					{
						"name": "sequence",
						"in": "path",
						"required": true,
						"schema": { "type": "string", "pattern": "^\\d+$" }
					}
				],
				"responses": {
					"200": {
						"description": "Ledger ingestion status",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LedgerIngestionStatus"
								}
							}
						}
					},
					"400": { "description": "Invalid ledger sequence" },
					"500": { "description": "Internal server error" }
				}
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"FullHistoryStatus": {
				"type": "object",
				"required": [
					"generatedAt",
					"status",
					"mode",
					"parsedLedgerCount",
					"earliestParsedLedger",
					"latestParsedLedger",
					"latestObservedAt",
					"sourceArchiveCount",
					"localTransactionIndexReady",
					"localOperationIndexReady",
					"localAssetIndexReady",
					"localContractIndexReady"
				],
				"properties": {
					"generatedAt": { "type": "string", "format": "date-time" },
					"status": {
						"type": "string",
						"enum": ["ok", "degraded", "unavailable"]
					},
					"mode": {
						"type": "string",
						"enum": ["archive_header_parser"]
					},
					"parsedLedgerCount": { "type": "integer", "minimum": 0 },
					"earliestParsedLedger": { "type": "string", "nullable": true },
					"latestParsedLedger": { "type": "string", "nullable": true },
					"latestObservedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"sourceArchiveCount": { "type": "integer", "minimum": 0 },
					"localTransactionIndexReady": { "type": "boolean" },
					"localOperationIndexReady": { "type": "boolean" },
					"localAssetIndexReady": { "type": "boolean" },
					"localContractIndexReady": { "type": "boolean" }
				}
			},
			"IngestionQueueSummary": {
				"type": "object",
				"required": [
					"doneJobs",
					"pendingJobs",
					"takenJobs",
					"latestJobUpdateAt"
				],
				"properties": {
					"doneJobs": { "type": "integer", "minimum": 0 },
					"pendingJobs": { "type": "integer", "minimum": 0 },
					"takenJobs": { "type": "integer", "minimum": 0 },
					"latestJobUpdateAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					}
				}
			},
			"IngestionStatus": {
				"allOf": [
					{ "$ref": "#/components/schemas/FullHistoryStatus" },
					{
						"type": "object",
						"required": ["queue"],
						"properties": {
							"queue": {
								"$ref": "#/components/schemas/IngestionQueueSummary"
							}
						}
					}
				]
			},
			"IndexingJob": {
				"type": "object",
				"required": [
					"concurrency",
					"fromLedger",
					"latestScannedLedger",
					"remoteId",
					"status",
					"toLedger",
					"updatedAt",
					"url"
				],
				"properties": {
					"concurrency": { "type": "integer", "nullable": true },
					"fromLedger": { "type": "string", "nullable": true },
					"latestScannedLedger": { "type": "string" },
					"remoteId": { "type": "string" },
					"status": { "type": "string", "enum": ["DONE", "PENDING", "TAKEN"] },
					"toLedger": { "type": "string", "nullable": true },
					"updatedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"url": { "type": "string" }
				}
			},
			"IndexingJobs": {
				"type": "object",
				"required": ["generatedAt", "jobs", "limit", "summary"],
				"properties": {
					"generatedAt": { "type": "string", "format": "date-time" },
					"jobs": {
						"type": "array",
						"items": { "$ref": "#/components/schemas/IndexingJob" }
					},
					"limit": { "type": "integer", "minimum": 1, "maximum": 100 },
					"summary": {
						"$ref": "#/components/schemas/IngestionQueueSummary"
					}
				}
			},
			"IndexingRange": {
				"type": "object",
				"required": [
					"archiveUrl",
					"earliestParsedLedger",
					"latestObservedAt",
					"latestParsedLedger",
					"parsedLedgerCount"
				],
				"properties": {
					"archiveUrl": { "type": "string" },
					"earliestParsedLedger": { "type": "string" },
					"latestObservedAt": { "type": "string", "format": "date-time" },
					"latestParsedLedger": { "type": "string" },
					"parsedLedgerCount": { "type": "integer", "minimum": 0 }
				}
			},
			"IndexingRanges": {
				"type": "object",
				"required": ["generatedAt", "limit", "ranges"],
				"properties": {
					"generatedAt": { "type": "string", "format": "date-time" },
					"limit": { "type": "integer", "minimum": 1, "maximum": 100 },
					"ranges": {
						"type": "array",
						"items": { "$ref": "#/components/schemas/IndexingRange" }
					}
				}
			},
			"LedgerIngestionHeader": {
				"type": "object",
				"required": [
					"bucketListHash",
					"ledgerHeaderHash",
					"protocolVersion",
					"sourceArchiveUrl",
					"transactionResultHash",
					"transactionSetHash"
				],
				"properties": {
					"bucketListHash": { "type": "string" },
					"ledgerHeaderHash": { "type": "string" },
					"protocolVersion": { "type": "integer" },
					"sourceArchiveUrl": { "type": "string" },
					"transactionResultHash": { "type": "string" },
					"transactionSetHash": { "type": "string" }
				}
			},
			"LedgerIngestionStatus": {
				"type": "object",
				"required": [
					"generatedAt",
					"header",
					"ledger",
					"parsedHeaderAvailable",
					"status"
				],
				"properties": {
					"generatedAt": { "type": "string", "format": "date-time" },
					"header": {
						"nullable": true,
						"allOf": [
							{ "$ref": "#/components/schemas/LedgerIngestionHeader" }
						]
					},
					"ledger": { "type": "string" },
					"parsedHeaderAvailable": { "type": "boolean" },
					"status": { "type": "string", "enum": ["parsed", "unparsed"] }
				}
			},
...
		}
	}
}